fix: Integration test
