+++ test/src/utils.rs
@@ -68,7 +68,8 @@ pub fn new_block_with_template(template: BlockTemplate) -> Block {
         .difficulty(template.difficulty.clone())
         .timestamp(template.current_time.0)
         .parent_hash(template.parent_hash)
-        .seal(Seal::new(rand::random(), Bytes::new()));
+        .seal(Seal::new(rand::random(), Bytes::new()))
+        .dao(template.dao.into_bytes());
 
     BlockBuilder::default()
         .uncles(template.uncles)
